pytest: Separate TSV test
